# perf(gif): reduce lzw decoding time