--- a/lv_conf_template.h
+++ b/lv_conf_template.h
@@ -618,6 +618,11 @@
 
 /*GIF decoder library*/
 #define LV_USE_GIF 0
+#if LV_USE_GIF
+/*GIF decoder accelerate*/
+#define LV_GIF_CACHE_LZW_CODE 0
+#endif
+
 
 /*Decode bin images to RAM*/
 #define LV_BIN_DECODER_RAM_LOAD 0