--- a/src/libs/gif/gifdec.c
+++ b/src/libs/gif/gifdec.c
@@ -23,9 +23,11 @@ typedef struct Table {
     Entry * entries;
 } Table;
 
+#if LV_GIF_CACHE_LZW_CODE
 #define LZW_MAXBITS                 12
 #define LZW_TABLE_SIZE              (1 << LZW_MAXBITS)
 #define LZW_CACHE_SIZE              (LZW_TABLE_SIZE * 4)
+#endif
 
 static gd_GIF  * gif_open(gd_GIF * gif);
 static bool f_gif_open(gd_GIF * gif, const void * path, bool is_file);
@@ -111,8 +113,11 @@ static gd_GIF * gif_open(gd_GIF * gif_base)
     /* Aspect Ratio */
     f_gif_read(gif_base, &aspect, 1);
     /* Create gd_GIF Structure. */
+#if LV_GIF_CACHE_LZW_CODE
     gif = lv_malloc(sizeof(gd_GIF) + 5 * width * height + LZW_CACHE_SIZE);
-
+    #else
+    gif = lv_malloc(sizeof(gd_GIF) + 5 * width * height);
+    #endif
     if(!gif) goto fail;
     memcpy(gif, gif_base, sizeof(gd_GIF));
     gif->width  = width;
@@ -129,7 +134,9 @@ static gd_GIF * gif_open(gd_GIF * gif_base)
         memset(gif->frame, gif->bgindex, gif->width * gif->height);
     }
     bgcolor = &gif->palette->colors[gif->bgindex * 3];
+    #if LV_GIF_CACHE_LZW_CODE
     gif->lzw_cache = gif->frame + width * height;
+    #endif
 
 #ifdef GIFDEC_FILL_BG
     GIFDEC_FILL_BG(gif->canvas, gif->width * gif->height, 1, gif->width * gif->height, bgcolor, 0xff);
@@ -310,6 +317,7 @@ get_key(gd_GIF *gif, int key_size, uint8_t *sub_len, uint8_t *shift, uint8_t *by
     return key;
 }
 
+#if LV_GIF_CACHE_LZW_CODE
 static int
 read_image_data(gd_GIF *gif, int interlace)
 {
@@ -450,6 +458,147 @@ read_image_data(gd_GIF *gif, int interlace)
     f_gif_seek(gif, end, LV_FS_SEEK_SET);
     return ret;
 }
+#else
+static Table *
+new_table(int key_size)
+{
+    int key;
+    int init_bulk = MAX(1 << (key_size + 1), 0x100);
+    Table * table = lv_malloc(sizeof(*table) + sizeof(Entry) * init_bulk);
+    if(table) {
+        table->bulk = init_bulk;
+        table->nentries = (1 << key_size) + 2;
+        table->entries = (Entry *) &table[1];
+        for(key = 0; key < (1 << key_size); key++)
+            table->entries[key] = (Entry) {
+            1, 0xFFF, key
+        };
+    }
+    return table;
+}
+
+/* Add table entry. Return value:
+ *  0 on success
+ *  +1 if key size must be incremented after this addition
+ *  -1 if could not realloc table */
+static int
+add_entry(Table ** tablep, uint16_t length, uint16_t prefix, uint8_t suffix)
+{
+    Table * table = *tablep;
+    if(table->nentries == table->bulk) {
+        table->bulk *= 2;
+        table = lv_realloc(table, sizeof(*table) + sizeof(Entry) * table->bulk);
+        if(!table) return -1;
+        table->entries = (Entry *) &table[1];
+        *tablep = table;
+    }
+    table->entries[table->nentries] = (Entry) {
+        length, prefix, suffix
+    };
+    table->nentries++;
+    if((table->nentries & (table->nentries - 1)) == 0)
+        return 1;
+    return 0;
+}
+
+/* Compute output index of y-th input line, in frame of height h. */
+static int
+interlaced_line_index(int h, int y)
+{
+    int p; /* number of lines in current pass */
+
+    p = (h - 1) / 8 + 1;
+    if(y < p)  /* pass 1 */
+        return y * 8;
+    y -= p;
+    p = (h - 5) / 8 + 1;
+    if(y < p)  /* pass 2 */
+        return y * 8 + 4;
+    y -= p;
+    p = (h - 3) / 4 + 1;
+    if(y < p)  /* pass 3 */
+        return y * 4 + 2;
+    y -= p;
+    /* pass 4 */
+    return y * 2 + 1;
+}
+
+/* Decompress image pixels.
+ * Return 0 on success or -1 on out-of-memory (w.r.t. LZW code table). */
+static int
+read_image_data(gd_GIF * gif, int interlace)
+{
+    uint8_t sub_len, shift, byte;
+    int init_key_size, key_size, table_is_full = 0;
+    int frm_off, frm_size, str_len = 0, i, p, x, y;
+    uint16_t key, clear, stop;
+    int ret;
+    Table * table;
+    Entry entry = {0};
+    size_t start, end;
+
+    f_gif_read(gif, &byte, 1);
+    key_size = (int) byte;
+    start = f_gif_seek(gif, 0, LV_FS_SEEK_CUR);
+    discard_sub_blocks(gif);
+    end = f_gif_seek(gif, 0, LV_FS_SEEK_CUR);
+    f_gif_seek(gif, start, LV_FS_SEEK_SET);
+    clear = 1 << key_size;
+    stop = clear + 1;
+    table = new_table(key_size);
+    key_size++;
+    init_key_size = key_size;
+    sub_len = shift = 0;
+    key = get_key(gif, key_size, &sub_len, &shift, &byte); /* clear code */
+    frm_off = 0;
+    ret = 0;
+    frm_size = gif->fw * gif->fh;
+    while(frm_off < frm_size) {
+        if(key == clear) {
+            key_size = init_key_size;
+            table->nentries = (1 << (key_size - 1)) + 2;
+            table_is_full = 0;
+        }
+        else if(!table_is_full) {
+            ret = add_entry(&table, str_len + 1, key, entry.suffix);
+            if(ret == -1) {
+                lv_free(table);
+                return -1;
+            }
+            if(table->nentries == 0x1000) {
+                ret = 0;
+                table_is_full = 1;
+            }
+        }
+        key = get_key(gif, key_size, &sub_len, &shift, &byte);
+        if(key == clear) continue;
+        if(key == stop || key == 0x1000) break;
+        if(ret == 1) key_size++;
+        entry = table->entries[key];
+        str_len = entry.length;
+        for(i = 0; i < str_len; i++) {
+            p = frm_off + entry.length - 1;
+            x = p % gif->fw;
+            y = p / gif->fw;
+            if(interlace)
+                y = interlaced_line_index((int) gif->fh, y);
+            gif->frame[(gif->fy + y) * gif->width + gif->fx + x] = entry.suffix;
+            if(entry.prefix == 0xFFF)
+                break;
+            else
+                entry = table->entries[entry.prefix];
+        }
+        frm_off += str_len;
+        if(key < table->nentries - 1 && !table_is_full)
+            table->entries[table->nentries - 1].suffix = entry.suffix;
+    }
+    lv_free(table);
+    if(key == stop) f_gif_read(gif, &sub_len, 1);  /* Must be zero! */
+    f_gif_seek(gif, end, LV_FS_SEEK_SET);
+    return 0;
+}
+
+#endif
 
 /* Read image.
  * Return 0 on success or -1 on out-of-memory (w.r.t. LZW code table). */