--- a/src/libs/gif/gifdec.h
+++ b/src/libs/gif/gifdec.h
@@ -43,6 +43,7 @@ typedef struct _gd_GIF {
     uint16_t fx, fy, fw, fh;
     uint8_t bgindex;
     uint8_t * canvas, * frame;
+    uint8_t *lzw_cache;
 } gd_GIF;
 
 gd_GIF * gd_open_gif_file(const char * fname);