--- a/src/lv_conf_internal.h
+++ b/src/lv_conf_internal.h
@@ -2054,7 +2054,7 @@
         #define LV_USE_GIF 0
     #endif
 #endif
-
+#if LV_USE_GIF
 /*GIF decoder accelerate*/
 #ifndef LV_GIF_CACHE_LZW_CODE
     #ifdef CONFIG_LV_GIF_CACHE_LZW_CODE
@@ -2063,6 +2063,8 @@
         #define LV_GIF_CACHE_LZW_CODE 0
     #endif
 #endif
+#endif
+
 
 /*Decode bin images to RAM*/
 #ifndef LV_BIN_DECODER_RAM_LOAD