@xaowang96 2023-11-30
> Cool!
> 
> Am I correct to assume that it has 16kB extra RAM usage per gif?

Yes, you are right. Actually