@xaowang96 2023-11-30
> How complicated would it be to add an `#if LV_GIF_CACHE_FRAME` (or so) option for it? 

That make sense.I will update it later.