@kisvegabor 2023-12-01
Thank you. Just a minor thing: I think `LV_GIF_CACHE_LZW_CODE` is too specific to the internals and something more generic would be easier to understand by the users.  E.g. `LV_GIF_CACHE_DECODE_DATA` or something similar?